getopt_long: hack to use global vars defined by unistd.h / lib_getopt.c
Using getopt_long from systemlib/getopt_long.h defines the optarg, optind
and optopt variables. These conflict with the ones declared in unistd.h
and defined in lib_getopt.c, which apperently is linked into the
application even if not #included.

To avoid having multiple versions of those globally defined variables,
we just let getopt_long use the versions declared in unistd.h. It is
necessary to have just one definition of those variables to be able to
reset optind and check optopt.